The books are now converted from .fb2.zip master. Also covers.
What we achieve by that is better look and feel of the output
document, namely the default font is no longer being reset to Times
New Roman. The book also have cover images; most of them are of
upscaled from an extremely poor thumbnail, so the quality is crap. But
I only care about the home page of kindle, and in there they look fine.